pyproject.toml
[project]
name = "steam-platform-stats"
version = "0.2.0"
description = "A tool to display user's Steam games statistics by platform"
requires-python = ">=3.12"
dependencies = [
    "argcomplete>=3.6.3",
    "prettytable>=3.17.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
]

[project.scripts]
steam-platform-stats = "steam_platform_stats.main:main"

[tool.setuptools.package-data]
"steam_platform_stats" = ["*.sh"]